slightly refactoring architecture
