+++ feeder/AirModel.cs
@@ -1,8 +1,9 @@
 namespace feeder
 {
+    using feeder.Binding;
     using System;
-    using System.Data.Entity;
     using System.ComponentModel.DataAnnotations.Schema;
+    using System.Data.Entity;
     using System.Linq;
 
     public partial class AirModel : DbContext
